aider: Added setInterestedSupplements method to Customer class.
# Aider chat conversation:

USER: The method setInterestedSupplements(List<String>) is undefined for the type Customer

ASSISTANT: It seems like the `Customer` class is missing the `setInterestedSupplements(List<String>)` method. We need to add this method to the `Customer` class. Here is the edit block for this change:

```java
src\Customer.java
<<<<<<< HEAD
    public List<String> getInterestedSupplements() {
        return this.interestedSupplements;
    }
=======
    public List<String> getInterestedSupplements() {
        return this.interestedSupplements;
    }

    public void setInterestedSupplements(List<String> interestedSupplements) {
        this.interestedSupplements = interestedSupplements;
    }
>>>>>>> updated
```

This change will add the `setInterestedSupplements(List<String>)` method to the `Customer` class, which allows us to update the list of supplements a customer is interested in.