aider: Added functionality to set or update a customer's payment method.
# Aider chat conversation:

USER: complete case 7

ASSISTANT: To complete case 7, we need to add the functionality to set or update a customer's payment method. This will involve changes in the `Main.java` file.

We will need to prompt the user for the type of payment method (Bank Account or Credit Card) and the details of the payment method (account number or card number). Then, we will create a new `PaymentMethod` object and set it as the customer's payment method.

Here are the changes needed:

```java
src\Main.java
<<<<<<< HEAD
                case "7":
                    // Set or update a customer's payment method
                    break;
=======
                case "7":
                    // Set or update a customer's payment method
                    email = cli.promptForEmail();
                    Customer customerToUpdatePayment = this.customers.stream().filter(c -> c.getEmail().equals(email)).findFirst().orElse(null);
                    if (customerToUpdatePayment != null) {
                        String paymentMethodType = cli.promptForPaymentMethodType();
                        if (paymentMethodType.equals("Bank Account")) {
                            String accountNumber = cli.promptForAccountNumber();
                            BankAccount bankAccount = new BankAccount(accountNumber);
                            customerToUpdatePayment.setPaymentMethod(bankAccount);
                        } else if (paymentMethodType.equals("Credit Card")) {
                            String cardNumber = cli.promptForCardNumber();
                            CreditCard creditCard = new CreditCard(cardNumber);
                            customerToUpdatePayment.setPaymentMethod(creditCard);
                        }
                    }
                    break;
>>>>>>> updated
```

Please note that this change assumes that the `Customer` class has a `setPaymentMethod(PaymentMethod paymentMethod)` method and the `CLI` class has `promptForPaymentMethodType()`, `promptForAccountNumber()`, and `promptForCardNumber()` methods. If these methods do not exist, they will need to be implemented.