aider: Set the weekly cost of the main magazine.
# Aider chat conversation:

USER: complete case 8

ASSISTANT: To complete case 8, we need to add the functionality to set the weekly cost of the main magazine. This will involve changes in the `Main.java` file.

We will need to prompt the user for the new weekly cost. Then, we will set the weekly cost of the magazine to the new cost.

Here are the changes needed:

```java
src\Main.java
<<<<<<< HEAD
                case "8":
                    // Set the weekly cost of the main magazine
                    break;
=======
                case "8":
                    // Set the weekly cost of the main magazine
                    double newWeeklyCost = cli.promptForCost();
                    this.magazine.setWeeklyCost(newWeeklyCost);
                    break;
>>>>>>> updated
```

Please note that this change assumes that the `Magazine` class has a `setWeeklyCost(double weeklyCost)` method and the `CLI` class has a `promptForCost()` method. If these methods do not exist, they will need to be implemented.