aider: Changed the `addSupplement` method in the `Magazine` class to accept a `Supplement` object instead of a `String`.
# Aider chat conversation:

USER: do those methods exist? if not add them

ASSISTANT: The `setWeeklyCost(double weeklyCost)` method exists in the `Magazine.java` file and the `promptForCost()` method exists in the `CLI.java` file.

However, if you want to add any new methods, please specify the methods and their functionalities.

USER: The method addSupplement(String) in the type Magazine is not applicable for the arguments (Supplement)

ASSISTANT: I see the issue. The `addSupplement` method in the `Magazine` class is expecting a `String` argument, but you're trying to pass a `Supplement` object to it.

We need to modify the `addSupplement` method in the `Magazine` class to accept a `Supplement` object instead of a `String`.

Here are the changes needed:

```java
src\Magazine.java
<<<<<<< HEAD
    public void addSupplement(String supplement) {
        this.supplements.add(supplement);
    }
=======
    public void addSupplement(Supplement supplement) {
        this.supplements.add(supplement);
    }
>>>>>>> updated
```

Please note that this change assumes that the `supplements` field in the `Magazine` class is a `List<Supplement>`. If it's not, we will need to change the type of the `supplements` field as well.