aider: Changed the `removeSupplement` method in the `Magazine` class to accept a `Supplement` object instead of a `String`.
# Aider chat conversation:

USER: The method removeSupplement(String) in the type Magazine is not applicable for the arguments (Supplement)

ASSISTANT: I see the issue. The `removeSupplement` method in the `Magazine` class is expecting a `String` argument, but you're trying to pass a `Supplement` object to it.

We need to modify the `removeSupplement` method in the `Magazine` class to accept a `Supplement` object instead of a `String`.

Here are the changes needed:

```java
src\Magazine.java
<<<<<<< HEAD
    public void removeSupplement(String supplement) {
        this.supplements.remove(supplement);
    }
=======
    public void removeSupplement(Supplement supplement) {
        this.supplements.remove(supplement);
    }
>>>>>>> updated
```

Please note that this change assumes that the `supplements` field in the `Magazine` class is a `List<Supplement>`. If it's not, we will need to change the type of the `supplements` field as well.