aider: Added getSupplementByName(String name) method to Magazine class.
# Aider chat conversation:

USER: getSupplementByName(String) is undefined for the type Magazine

ASSISTANT: It seems like you're trying to access a supplement by its name from the `Magazine` class, but there's no such method defined in the `Magazine` class. We can add a method `getSupplementByName(String name)` to the `Magazine` class to solve this issue.

Here's how we can do it:

1. Modify the `src\Magazine.java` file to add the `getSupplementByName(String name)` method.

Here's the *edit block* for the change:

```java
src\Magazine.java
<<<<<<< HEAD
    public void modifySupplement(String oldSupplement, String newSupplement) {
        int index = this.supplements.indexOf(oldSupplement);
        if (index != -1) {
            this.supplements.set(index, newSupplement);
        }
    }
=======
    public Supplement getSupplementByName(String name) {
        return this.supplements.stream()
            .filter(supplement -> supplement.getName().equals(name))
            .findFirst()
            .orElse(null);
    }

    public void modifySupplement(String oldSupplement, String newSupplement) {
        int index = this.supplements.indexOf(oldSupplement);
        if (index != -1) {
            this.supplements.set(index, newSupplement);
        }
    }
>>>>>>> updated
```

This new method `getSupplementByName(String name)` will return the first supplement that matches the given name, or `null` if no such supplement is found.