aider: Added method getSupplementByName(String name) to the Magazine class.
# Aider chat conversation:

USER: The method getSupplementByName(String) is undefined for the type Magazine

ASSISTANT: I see that the `getSupplementByName(String)` method is indeed missing from the `Magazine` class. I can add this method for you. This method will return a `Supplement` object that matches the provided name. If no match is found, it will return `null`.

Here is the change:

```java
src\Magazine.java
<<<<<<< HEAD
    public void modifySupplement(String oldSupplement, String newSupplement) {
        int index = this.supplements.indexOf(oldSupplement);
        if (index != -1) {
            this.supplements.set(index, newSupplement);
        }
    }
=======
    public void modifySupplement(String oldSupplement, String newSupplement) {
        int index = this.supplements.indexOf(oldSupplement);
        if (index != -1) {
            this.supplements.set(index, newSupplement);
        }
    }

    public Supplement getSupplementByName(String name) {
        for (Supplement supplement : this.supplements) {
            if (supplement.getName().equals(name)) {
                return supplement;
            }
        }
        return null;
    }
>>>>>>> updated
```

This new method `getSupplementByName(String name)` will iterate over the list of supplements and return the first supplement that matches the provided name. If no match is found, it will return `null`.